requirements.txt
-i https://pypi.org/simple/
amqp==2.3.2
billiard==3.5.0.4
psycopg2-binary>=2.8
celery==4.2.1
Django==3.1.1
django-celery-beat==1.1.1
django-celery-results==1.0.1
kombu==4.2.1
psycopg2==2.8.4
pytz==2018.5
vine==1.1.4
typing~=3.7.4.3
requests==2.24.0
PyYAML==5.3.1
django-environ==0.4.5
redis==2.10.6
waitress==1.2.1
gunicorn==20.0.4